adjust comment & style for main.c
This version shows a continuous interrupt-based ADC example. There is a
counter "count_interrupt" sets the bitmap of 4 LEDs, initialized to 10.
During execution, the counter decreases, and LEDs show 1010->1001->1000
->0111->...->0001->0000, where 0/1 represents emitted/lighted.

The program opens EOC interrupt when it is ready to handle EOC. It waits
for a change of counter after enabling EOC, keeping the track of exact one
execution of IRQ handler.